refactor(cloud): extract viewer session adapter
